+++ tango-dataview/src/save.rs
@@ -190,8 +190,8 @@ pub trait NavicustViewMut<'a> {
 }
 
 pub trait AutoBattleDataView<'a> {
-    fn chip_use_count(&self, id: usize) -> Option<u16>;
-    fn secondary_chip_use_count(&self, id: usize) -> Option<u16>;
+    fn chip_use_count(&self, id: usize) -> Option<usize>;
+    fn secondary_chip_use_count(&self, id: usize) -> Option<usize>;
     fn materialized(&self) -> crate::abd::MaterializedAutoBattleData;
 }
